Refactor code to clean and parse hyperlinks in url-scraperThe commit refactors the code in url-scraper to clean text and parse hyperlinks using separate functions. This improves readability and maintainability of the code
